Improve how filenames are picked up
The previous version of the stile sheet relied on the content of the <title> tag to report the file name (i.e. ch1.xhtml). This was a working choice previous the introduction of the package 'obp-epub-fixup' which replaces the content of <title> with the more informative string contained in <h1>.

The mod in this commit relies on the base-uri() function of xsl to report the path of the file being processed, and then basename and file extension are stripped out.